requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
cmocean==3.0.3
colorcet==3.0.1
colorlog==6.8.2
comm==0.2.0
contourpy==1.1.0
cv==1.0.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
edt==2.3.1
einops==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastcore==1.5.29
fastjsonschema==2.19.0
filelock==3.12.3
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
imageio==2.33.0
importlib-metadata==7.0.0
importlib-resources==6.0.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.4.4
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MedPy==0.4.0
meshio==5.3.4
mistune==3.0.2
monai==1.2.0
mpmath==1.3.0
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.1.0
nptyping==2.5.0
opencv-python==4.9.0.80
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
param==2.0.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pooch==1.8.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyct==0.5.0
pydicom==2.4.3
Pygments==2.17.2
pynrrd==1.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.2.0.post0
pytz==2023.3.post1
pyvista==0.43.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
scooby==0.9.2
seaborn==0.13.0
Send2Trash==1.8.2
sentry-sdk==1.38.0
setproctitle==1.3.3
simpervisor==1.0.0
SimpleITK==2.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
#torch==2.0.1+cu118
#torchaudio==2.0.2+cu118
#torchmetrics==1.3.1
#torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trame==3.3.0
trame-client==2.13.0
trame-server==2.12.1
trame-vtk==2.6.2
trame-vuetify==2.3.1
triton==2.0.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
vtk==9.3.0
wandb==0.16.3
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wslink==1.12.4
yarl==1.9.4
zipp==3.16.2
